availability-management/list-availability-servicePy/requirements.txt
Flask
Flask-Swagger-UI
pymysql
pika